requirements.txt
numpy
matplotlib
pyyaml
ortools
pytest
pyscipopt
gurobipy
cuopt_sh_client